requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
sounddevice>=0.4.6
noisereduce>=3.0.0
pyttsx3>=2.90
streamlit>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
